refactor: 부서 서비스 분리 및 UserService에서 호출하도록 변경
