+++ src/main/java/com/example/bookmanagementsystembo/department/controller/DepartmentController.java
@@ -1,0 +1,4 @@
+package com.example.bookmanagementsystembo.department.controller;
+
+public class DepartmentController {
+}
