+++ src/main/java/com/example/bookmanagementsystembo/department/dto/DepartmentInfo.java
@@ -1,0 +1,7 @@
+package com.example.bookmanagementsystembo.department.dto;
+
+public record DepartmentInfo(String name) {
+    public static DepartmentInfo of(String name) {
+        return new DepartmentInfo(name);
+    }
+}
