+++ src/main/java/com/example/bookmanagementsystembo/user/domain/service/UserService.java
@@ -1,11 +1,11 @@
 package com.example.bookmanagementsystembo.user.domain.service;
 
+import com.example.bookmanagementsystembo.department.dto.DepartmentInfo;
+import com.example.bookmanagementsystembo.department.service.DepartmentService;
 import com.example.bookmanagementsystembo.exception.CoreException;
 import com.example.bookmanagementsystembo.exception.ErrorType;
 import com.example.bookmanagementsystembo.user.domain.dto.UserInfo;
-import com.example.bookmanagementsystembo.user.domain.entity.Department;
 import com.example.bookmanagementsystembo.user.domain.entity.Users;
-import com.example.bookmanagementsystembo.user.infra.DepartmentRepository;
 import com.example.bookmanagementsystembo.user.infra.UserRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
@@ -16,15 +16,14 @@ public class UserService {
 
     private final UserRepository userRepository;
 
-    private final DepartmentRepository departmentRepository;
+    private final DepartmentService departmentService;
 
     public UserInfo findByUser(Long userId) {
         Users user = userRepository.findById(userId)
                 .orElseThrow(() -> new CoreException(ErrorType.USER_NOT_FOUND, userId));
 
-        Department department = departmentRepository.findById(user.getDepartmentId())
-                .orElseThrow(() -> new CoreException(ErrorType.DEPARTMENT_NOT_FOUND, user.getDepartmentId()));
+        DepartmentInfo department = departmentService.findById(user.getDepartmentId());
 
-        return UserInfo.of(user.getName(), user.getEmail(), department.getName());
+        return UserInfo.of(user.getName(), user.getEmail(), department.name());
     }
 }
